requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Andrew\Documents\Lagann\Lagann\connection.py: 4,5,6
# C:\Users\Andrew\Documents\Lagann\Lagann\lagcerts.py: 2,3,4
cryptography == 36.0.1

# C:\Users\Andrew\Documents\Lagann\Lagann\connection.py: 3
# C:\Users\Andrew\Documents\Lagann\Lagann\lagcerts.py: 1
pyOpenSSL == 21.0.0

# C:\Users\Andrew\Documents\Lagann\Lagann\Lagann\Lagann\__init__.py: 8
# C:\Users\Andrew\Documents\Lagann\Lagann\view.py: 1
py_cui == 0.1.4

# C:\Users\Andrew\Documents\Lagann\Lagann\Lagann\setup.py: 1
setuptools == 50.3.0
